Rename uart datagrams to serial datagrams
The rationale behind this move is that those datagrams can be used on
other types of serial link, such as TCP streams.
